Renamed item actions to actions. Moved enabled columns into actions. Some cleanup and bug fixes.
